# api/rest/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Swagger API for Kafka Test.",
        "title": "Kafka Test Swagger API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://github.com/YEgorLu/kafka-test/blob/master/LICENSE"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/messages": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "messages"
                ],
                "summary": "Creates message and puts it in queue for later processing",
                "parameters": [
                    {
                        "description": "Your message",
                        "name": "value",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "{\"value\": \"your message\"}"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.AddResponse"
                        }
                    }
                }
            }
        },
        "/messages/stats": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "messages"
                ],
                "summary": "Returns overall and processed messages count",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.StatsResponse"
                        }
                    }
                }
            }
        },
        "/messages/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "messages"
                ],
                "summary": "Returns message",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Message ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GetOneResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.AddResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "models.GetOneResponse": {
            "type": "object",
            "properties": {
                "dateCreate": {
                    "type": "string"
                },
                "dateProcess": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "processed": {
                    "type": "boolean"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "models.StatsResponse": {
            "type": "object",
            "properties": {
                "countAll": {
                    "type": "integer"
                },
                "countProcessed": {
                    "type": "integer"
                }
            }
        }
    }
}